requirements.txt
fastapi
python-multipart
pillow
numpy==1.19.2
torch==1.6.0
torchvision==0.7.0
opencv-python
uvicorn
pydantic-settings